able to generate terrain with textures :)(finally)

+++ Healthcare test/VR/Session.cs
@@ -197,18 +197,41 @@ namespace Healthcare_test.VR
                 }
                 if (jsonData.data.data.id == "scene/node/find")
                 {
+                    //System.Diagnostics.Debug.WriteLine(information);
                     string uuid = jsonData.data.data.data[0].uuid;
                     System.Diagnostics.Debug.WriteLine(uuid + " test");
-                    Send(JsonConvert.SerializeObject(Commands.DeleteNode(gui.tunnel, uuid)));
+                    System.Diagnostics.Debug.WriteLine(information);
+                    System.Diagnostics.Debug.WriteLine("naam is " + (string)jsonData.data.data.data[0].name);
+                    if (jsonData.data.data.data[0].name == "GroundPlane")
+                    {
+                        System.Diagnostics.Debug.WriteLine("dit is de groundplane");
+                        Send(JsonConvert.SerializeObject(Commands.DeleteNode(gui.tunnel, uuid)));
+                    }
+                    if (jsonData.data.data.data[0].name == "name")
+                    {
+                        System.Diagnostics.Debug.WriteLine("texture proberen toe te voegen");
+                        SetupTexturesTerrain(uuid);
+                        
+                    }
+
                 }
                 else
                 {
-
-                    System.Diagnostics.Debug.WriteLine("Else: \r\n" + information);
+                     System.Diagnostics.Debug.WriteLine("Else: \r\n" + information);
                 }
             }
 
 
+        }
+
+        private void SetupTexturesTerrain(string uuid)
+        {
+            Send(JsonConvert.SerializeObject(Commands.addTextureTerrain(gui.tunnel, uuid, "desert_wet_n.jpg",0,2,1 ))); 
+            Send(JsonConvert.SerializeObject(Commands.addTextureTerrain(gui.tunnel, uuid, "grass_ground2y_d.jpg", 2, 5, 0)));
+            Send(JsonConvert.SerializeObject(Commands.addTextureTerrain(gui.tunnel, uuid, "snow_mud_d.jpg", 5, 15, 0)));
+            Send(JsonConvert.SerializeObject(Commands.addTextureTerrain(gui.tunnel, uuid, "mntn_x2_d.jpg", 15, 28, 0)));
+            Send(JsonConvert.SerializeObject(Commands.addTextureTerrain(gui.tunnel, uuid, "snow_rough_s.jpg", 28, 40, 0)));
+            Send(JsonConvert.SerializeObject(Commands.addTextureTerrain(gui.tunnel, uuid, "snow2ice_d.jpg", 40, 100, 0)));
         }
 
         public void ProcessSessionList(dynamic information)
